[업데이트]
대시보드 그래프 생성

게시판관리 삭제 , 선택삭제 , 추가 수정

수정은 아직

+++ src/component/Manager/router/User/List.jsx
@@ -340,10 +340,12 @@ export default function List() {
                 <Button 
                     color={"color02"}
                     onClick={()=>{
-                        dispatch(multipleDeleteAction(chkItem));
-                        if( searchParams.get('page') >= totalPage){
-                            searchParams.set('page',last-1);
-                            setSearchParams(searchParams);
+                        if(window.confirm('삭제 하시겠습니까?')){
+                            dispatch(multipleDeleteAction(chkItem));
+                            if( searchParams.get('page') >= totalPage){
+                                searchParams.set('page',last-1);
+                                setSearchParams(searchParams);
+                            }
                         }
                     }}
                 >선택 삭제</Button>
